+++ tests/s/test_schktr.c
@@ -278,7 +278,7 @@ static void test_standard(void** state)
 
             if (result[0] >= THRESH) {
                 print_message("TEST 1 failed: n=%d, imat=%d, uplo=%s, diag=%c, nb=%d, resid=%.3e\n",
-                             n, imat, uplo, diag, NBVAL[inb], result[0]);
+                             n, imat, uplo, diag, NBVAL[inb], (double)result[0]);
             }
             assert_residual_ok(result[0]);
         }
@@ -332,7 +332,7 @@ static void test_standard(void** state)
                 for (int k = 1; k < 6; k++) {
                     if (result[k] >= THRESH) {
                         print_message("TEST %d failed: n=%d, imat=%d, uplo=%s, trans=%s, nrhs=%d, resid=%.3e\n",
-                                     k + 1, n, imat, uplo, trans, nrhs, result[k]);
+                                     k + 1, n, imat, uplo, trans, nrhs, (double)result[k]);
                     }
                     assert_residual_ok(result[k]);
                 }
@@ -361,7 +361,7 @@ static void test_standard(void** state)
 
                 if (result[6] >= THRESH) {
                     print_message("TEST 7 failed: n=%d, imat=%d, uplo=%s, norm=%s, resid=%.3e\n",
-                                 n, imat, uplo, norm, result[6]);
+                                 n, imat, uplo, norm, (double)result[6]);
                 }
                 assert_residual_ok(result[6]);
             }
@@ -467,15 +467,15 @@ static void test_latrs(void** state)
         /* Check results */
         if (result[7] >= THRESH) {
             print_message("TEST 8 failed: n=%d, imat=%d, uplo=%s, trans=%s, resid=%.3e\n",
-                         n, imat, uplo, trans, result[7]);
+                         n, imat, uplo, trans, (double)result[7]);
         }
         if (result[8] >= THRESH) {
             print_message("TEST 9 failed: n=%d, imat=%d, uplo=%s, trans=%s, resid=%.3e\n",
-                         n, imat, uplo, trans, result[8]);
+                         n, imat, uplo, trans, (double)result[8]);
         }
         if (result[9] >= THRESH) {
             print_message("TEST 10 failed: n=%d, imat=%d, uplo=%s, trans=%s, resid=%.3e\n",
-                         n, imat, uplo, trans, result[9]);
+                         n, imat, uplo, trans, (double)result[9]);
         }
         assert_residual_ok(result[7]);
         assert_residual_ok(result[8]);
